feat: Update client.py and data_node.py to include file download functionality
